# Add RB_UNLIKELY hint for tracing log