I think that `Yes == pi->options.trace` will be false in most cases.
If the compiler supports it, the UNLIKELY hint might speed up the conditional branching slightly.


−       | before   | after   | result
--       | --       | --      | --
Oj.load  | 374.821k |383.001k | 1.022x

### Environment
- Linux
  - Manjaro Linux x86_64
  - Kernel: 5.18.0-1-rt11-MANJARO 
  - AMD Ryzen 7 5700G
  - gcc version 12.1.0
  - Ruby 3.1.2

### Before
```
Warming up --------------------------------------
             Oj.load    37.784k i/100ms
Calculating -------------------------------------
             Oj.load    374.821k (± 0.5%) i/s -      7.519M in  20.060807s
```

### After
```
Warming up --------------------------------------
             Oj.load    38.599k i/100ms
Calculating -------------------------------------
             Oj.load    383.001k (± 0.5%) i/s -      7.681M in  20.055707s
```

### Test code
```ruby
require 'benchmark/ips'
require 'oj'

json =<<-EOF
{
  "$id": "https://example.com/person.schema.json",
  "$schema": "https://json-schema.org/draft/2020-12/schema",
  "title": "Person",
  "type": "object",
  "properties": {
    "firstName": {
      "type": "string",
      "description": "The person's first name."
    },
    "lastName": {
      "type": "string",
      "description": "The person's last name."
    },
    "age": {
      "description": "Age in years which must be equal to or greater than zero.",
      "type": "integer",
      "minimum": 0
    }
  }
}
EOF

Benchmark.ips do |x|
  x.warmup = 10
  x.time = 20

  x.report('Oj.load') { Oj.load(json) }
end
```